+++ Integrations/PaloAltoNetworks_Traps/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [19.10.1] - 2019-10-15
+Updated integration category to *Endpoint*.
+
+
 ## [19.10.0] - 2019-10-03
 #### New Integration
 Endpoint protection and response stops threats on endpoints and coordinates enforcement with network and cloud security to prevent successful cyberattacks.