+++ Integrations/ServiceNow/CHANGELOG.md
@@ -1,1 +1,8 @@
-Fixed an issue with the **servicenow-upload-file*** command when the uploaded file is an info file.
+## [Unreleased]
+
+
+## [19.10.1] - 2019-10-15
+  -
+
+## [19.10.0] - 2019-10-03
+  - Fixed an issue with the **servicenow-upload-file*** command when the uploaded file is an info file.